+++ python/src/scipp/_variable.py
@@ -421,59 +421,5 @@ def arange(dim: str,
                  dtype=dtype)
 
 
-def random(*,
-           dims: _Sequence[str] = None,
-           shape: _Sequence[int] = None,
-           sizes: dict = None,
-           unit: _Union[_cpp.Unit, str] = _cpp.units.dimensionless,
-           variances: bool = None) -> _cpp.Variable:
-    """Constructs a :class:`Variable` with random values between 0 and 1 with
-    given dimension labels and shape.
-    The dims and shape can also be specified using a sizes dict.
-    Optionally can add randomly initialized variances.
-    The dtype returned is always float64.
-
-    :seealso: :py:func:`scipp.zeros` :py:func:`scipp.ones`
-              :py:func:`scipp.random_like`
-
-    :param dims: Optional (if sizes is specified), dimension labels.
-    :param shape: Optional (if sizes is specified), dimension sizes.
-    :param sizes: Optional, imension label to size map.
-    :param unit: Optional, unit of contents. Default=dimensionless
-    :param variances: Optional, boolean flag, if True includes variances
-      initialized with random values. Default=False
-    """
-    params = _parse_dims_shape_sizes(dims, shape, sizes)
-
-    _variances = None
-    if variances is not None and variances is not False:
-        _variances = _np.random.random(params["shape"])
-
-    return array(dims=params["dims"],
-                 values=_np.random.random(params["shape"]),
-                 variances=_variances,
-                 unit=unit,
-                 dtype=_cpp.dtype.float64)
-
-
-def random_like(var: _cpp.Variable) -> _cpp.Variable:
-    """Constructs a :class:`Variable` with the same dims, shape, unit
-    as the input variable, but with all values initialized to a random number
-    between 0 and 1.
-    If the input has variances, the output will be initialized with random
-    variances.
-    The dtype returned is always float64.
-
-    :seealso: :py:func:`scipp.random` :py:func:`scipp.zeros_like`
-              :py:func:`scipp.ones_like`
-
-    :param var: Input variable.
-    """
-    return random(dims=var.dims,
-                  shape=var.shape,
-                  unit=var.unit,
-                  variances=var.variances is not None)
-
-
 # Wrapper to make datetime usable without importing numpy manually.
 datetime64 = _np.datetime64